Add :features:stopwatch module
