--- app/src/main/java/com/example/timicompose/stopwatch/domain/StopwatchStateHolder.kt
@@ -1,33 +1,0 @@
-package com.example.timicompose.stopwatch.domain
-
-import com.akjaw.core.common.domain.model.toTimestampMilliseconds
-import com.example.timicompose.stopwatch.domain.model.StopwatchState
-import com.example.timicompose.stopwatch.domain.utilities.ElapsedTimeCalculator
-import com.example.timicompose.stopwatch.domain.utilities.TimestampMillisecondsFormatter
-
-class StopwatchStateHolder(
-    private val stopwatchStateCalculator: StopwatchStateCalculator,
-    private val elapsedTimeCalculator: ElapsedTimeCalculator,
-    private val timestampMillisecondsFormatter: TimestampMillisecondsFormatter,
-) {
-
-    //TODO does this have to be thread-safe?
-    var currentState: StopwatchState = StopwatchState.Paused(0.toTimestampMilliseconds())
-        private set
-
-    fun start() {
-        currentState = stopwatchStateCalculator.calculateRunningState(currentState)
-    }
-
-    fun pause() {
-        currentState = stopwatchStateCalculator.calculatePausedState(currentState)
-    }
-
-    fun getStringTimeRepresentation(): String {
-        val elapsedTime = when (val currentState = currentState) {
-            is StopwatchState.Paused -> currentState.elapsedTime
-            is StopwatchState.Running -> elapsedTimeCalculator.calculate(currentState)
-        }
-        return timestampMillisecondsFormatter.format(elapsedTime)
-    }
-}
