--- app/src/main/java/com/example/timicompose/stopwatch/domain/model/StopwatchState.kt
@@ -1,15 +1,0 @@
-package com.example.timicompose.stopwatch.domain.model
-
-import com.akjaw.core.common.domain.model.TimestampMilliseconds
-
-sealed class StopwatchState {
-    data class Running(
-        val startTime: TimestampMilliseconds,
-        val elapsedTime: TimestampMilliseconds
-    ) : StopwatchState()
-    data class Paused(
-        val elapsedTime: TimestampMilliseconds
-    ) : StopwatchState()
-    //TODO Stopped?
-//        data class Stopped()
-}
