+++ build.gradle
@@ -17,6 +17,12 @@ buildscript {
     }
 }
 
+allprojects {
+    tasks.withType(Test) {
+        useJUnitPlatform()
+    }
+}
+
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
