+++ features/stopwatch/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'kotlin-android'
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
@@ -10,13 +10,13 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.timicompose"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -31,7 +31,6 @@ android {
     }
     kotlinOptions {
         jvmTarget = '1.8'
-        useIR = true
     }
     buildFeatures {
         compose true
@@ -40,15 +39,10 @@ android {
         kotlinCompilerExtensionVersion compose_version
         kotlinCompilerVersion '1.4.30'
     }
-    hilt {
-        enableExperimentalClasspathAggregation = true
-    }
 }
 
 dependencies {
-
     implementation project(":core")
-    implementation project(":features:stopwatch")
 
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
@@ -60,7 +54,6 @@ dependencies {
     implementation "androidx.navigation:navigation-compose:1.0.0-alpha09"
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.0-alpha06'
     implementation 'androidx.activity:activity-compose:1.3.0-alpha02'
-    implementation "com.soywiz.korlibs.klock:klock:2.0.7"
     implementation "androidx.lifecycle:lifecycle-viewmodel-compose:1.0.0-alpha03"
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.3'
 
@@ -73,12 +66,4 @@ dependencies {
     testImplementation "io.mockk:mockk:1.11.0"
     testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.4.3'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.6.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-}
-
-kotlin {
-    sourceSets.all {
-        languageSettings.enableLanguageFeature('InlineClasses')
-    }
 }
