+++ features/stopwatch/src/main/java/com/akjaw/stopwatch/domain/StopwatchStateCalculator.kt
@@ -1,0 +1,32 @@
+package com.akjaw.stopwatch.domain
+
+import com.akjaw.core.common.domain.TimestampProvider
+import com.akjaw.stopwatch.domain.model.StopwatchState
+import com.akjaw.stopwatch.domain.utilities.ElapsedTimeCalculator
+import javax.inject.Inject
+
+class StopwatchStateCalculator @Inject constructor(
+    private val timestampProvider: TimestampProvider,
+    private val elapsedTimeCalculator: ElapsedTimeCalculator,
+) {
+
+    fun calculateRunningState(oldState: StopwatchState): StopwatchState.Running =
+        when (oldState) {
+            is StopwatchState.Running -> oldState
+            is StopwatchState.Paused -> {
+                StopwatchState.Running(
+                    startTime = timestampProvider.getMilliseconds(),
+                    elapsedTime = oldState.elapsedTime
+                )
+            }
+        }
+
+    fun calculatePausedState(oldState: StopwatchState): StopwatchState.Paused =
+        when (oldState) {
+            is StopwatchState.Running -> {
+                val elapsedTime = elapsedTimeCalculator.calculate(oldState)
+                StopwatchState.Paused(elapsedTime = elapsedTime)
+            }
+            is StopwatchState.Paused -> oldState
+        }
+}
