+++ features/stopwatch/src/main/java/com/akjaw/stopwatch/domain/model/StopwatchState.kt
@@ -1,0 +1,15 @@
+package com.akjaw.stopwatch.domain.model
+
+import com.akjaw.core.common.domain.model.TimestampMilliseconds
+
+sealed class StopwatchState {
+    data class Running(
+        val startTime: TimestampMilliseconds,
+        val elapsedTime: TimestampMilliseconds
+    ) : StopwatchState()
+    data class Paused(
+        val elapsedTime: TimestampMilliseconds
+    ) : StopwatchState()
+    //TODO Stopped?
+//        data class Stopped()
+}
