+++ features/stopwatch/src/main/java/com/akjaw/stopwatch/presentation/StopwatchViewModel.kt
@@ -1,0 +1,43 @@
+package com.akjaw.stopwatch.presentation
+
+import androidx.lifecycle.ViewModel
+import com.akjaw.core.common.domain.data.TaskRepository
+import com.akjaw.core.common.domain.model.Task
+import com.akjaw.stopwatch.domain.StopwatchListOrchestrator
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.combineTransform
+import javax.inject.Inject
+
+@HiltViewModel
+class StopwatchViewModel @Inject constructor(
+    private val taskRepository: TaskRepository,
+    private val stopwatchListOrchestrator: StopwatchListOrchestrator,
+) : ViewModel() {
+
+    val stopwatches: StateFlow<Map<Task, String>> = stopwatchListOrchestrator.ticker
+
+    val availableTasks: Flow<List<Task>> = combineTransform(
+        MutableStateFlow(taskRepository.tasks),
+        stopwatches
+    ) { tasks, stopwatches ->
+        val availableTasks = tasks.filterNot { task ->
+            stopwatches.containsKey(task)
+        }
+        emit(availableTasks)
+    }
+
+    fun start(task: Task) = stopwatchListOrchestrator.start(task)
+
+    fun pause(task: Task) = stopwatchListOrchestrator.pause(task)
+
+    fun stop(task: Task) = stopwatchListOrchestrator.stop(task)
+
+    // TODO is this correct?
+    override fun onCleared() {
+        super.onCleared()
+        stopwatchListOrchestrator.destroy()
+    }
+}
